[REFACTOR #41]: domain 객체 수정 (#43)
* refactor: domain 객체 수정

* refactor: domain 객체 수정 - 타입 및 외래키 관계

+++ src/main/java/com/payper/benefit/domain/BenefitLimit.java
@@ -10,15 +10,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class BenefitLimit {
-    private Long benefitLimitId;
-
-    private Long benefitId;
-
-    private Integer limitCountPerDay;
-
-    private Integer limitCountPerMonth;
-
-    private Integer limitCountPerYear;
-
-    private Integer limitAmountPerPay;
+    private Integer benefitLimitId;
+    private Integer benefitId;
+    private Long limitCountPerDay;
+    private Long limitCountPerMonth;
+    private Long limitCountPerYear;
+    private Long limitAmountPerPay;
 }
